Fix handling JSON response to not panic
If the response from HTTP is not a valid JSON i.e. another
response as "invalid API_KEY" or "too many requests that not allowed"
or any other responses that will make serde_json to cause panic.

This is not an issue in Serde-JSON crate rather in handling the
right response to be passed into serde_json.

Signed-off-by: AhmedNGaber <ahmed.n.gaber@gmail.com>